This commit will move the unit test of transition to test/unit/transitions directory
