+++ shared-libs/transitions/test/unit/transitions/default_responses.js
@@ -1,8 +1,8 @@
 const sinon = require('sinon');
 const assert = require('chai').assert;
-const config = require('../../src/config');
-const messages = require('../../src/lib/messages');
-const transition = require('../../src/transitions/default_responses');
+const config = require('../../../src/config');
+const messages = require('../../../src/lib/messages');
+const transition = require('../../../src/transitions/default_responses');
 
 describe('default responses', () => {
   afterEach(() => sinon.restore());
