+++ shared-libs/transitions/test/unit/transitions/update_sent_by.js
@@ -1,7 +1,7 @@
 const sinon = require('sinon');
 const assert = require('chai').assert;
-const db = require('../../src/db');
-const transition = require('../../src/transitions/update_sent_by');
+const db = require('../../../src/db');
+const transition = require('../../../src/transitions/update_sent_by');
 
 describe('update sent by', () => {
   beforeEach(() => { process.env.TEST_ENV = true; });
